+++ src/main/java/org/flag4j/linalg/ops/dispatch/ReDeMatVecMultDispatcher.java
@@ -139,7 +139,7 @@ public final class ReDeMatVecMultDispatcher extends BiTensorOpDispatcher<Matrix,
      * @return The matrix product of {@code a} and {@code b}.
      */
     public static Vector dispatch(Matrix a, Vector b) {
-        int totalEntries = a.dataLength() + b.dataLength();
+        long totalEntries = (long) a.dataLength() + b.dataLength();
 
         // Return as fast as possible for "small-enough" matrices.
         if(totalEntries < SML_THRESH) {
@@ -167,7 +167,7 @@ public final class ReDeMatVecMultDispatcher extends BiTensorOpDispatcher<Matrix,
     protected BiFunction<Matrix, Vector, Vector> getFunc(Shape aShape, Shape bShape, int data1Length, int data2Length) {
         int m = aShape.get(0);
         int n = aShape.get(1);
-        double aspectRatio = Math.max(m, n) / Math.min(m, n);
+        double aspectRatio = (double) Math.max(m, n) / Math.min(m, n);
 
         if(aspectRatio <= 4*ASPECT_THRESH) {
             // The matrix is approximately square.
